requirements.txt
chainerrl
chainer
numpy
click
Pillow
tensorboard-chainer
matplotlib
